Add files via upload
I thought it would be interesting to see what the "most successful" actors are outside of this dataset and found one that merges nicely. Also for fun thought it would be cool to compare it with top100 most attractive actors. A lot more to be done, this will probably just be a simple section.